move the core to a module
