+++ .idea/.name
@@ -1,1 +1,1 @@
-LazyCountAnimation
+Lazy Count Animation
