+++ lazycountanimation/src/main/AndroidManifest.xml
@@ -1,0 +1,3 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.github.frog1014.lazycountanimation">
+</manifest>
