+++ settings.gradle
@@ -1,2 +1,3 @@
+include ':lazycountanimation'
 include ':app'
-rootProject.name = "LazyCountAnimation"
+rootProject.name = "Lazy Count Animation"
